Restructuring the AuthController and RegisterFormRequest
